Refactor map component to use props
