+++ src/pages/requests/index.tsx
@@ -8,6 +8,10 @@ import { Header } from "../../others/components/Header";
 import { Main } from "../../others/components/Main";
 import { Sidebar } from "../../others/components/Sidebar";
 import { MultiTab } from "../../others/components/MultiTab";
+import { Layer, Source } from "react-map-gl";
+import { layerStyle } from "../../others/components/map/CircleLayerStyle";
+import { aidRequestsFixture } from "../../others/fixtures/request.fixture";
+import { FeatureCollection, GeoJsonProperties, Geometry } from "geojson";
 
 export function Requests() {
   const { t } = useTranslation();
@@ -16,6 +20,11 @@ export function Requests() {
   if (!cities) {
     return <Layout header={<Header />}>{/* <Loader /> */}</Layout>;
   }
+
+  const geojson: FeatureCollection<Geometry, GeoJsonProperties> = {
+    type: "FeatureCollection",
+    features: aidRequestsFixture,
+  };
 
   return (
     <Layout header={<Header />}>
@@ -26,7 +35,13 @@ export function Requests() {
           </Sidebar>
         }
       >
-        <Map />
+        <Map
+          sourceWithLayer={
+            <Source id="ukr_water_needs" type="geojson" data={geojson}>
+              <Layer {...layerStyle} />
+            </Source>
+          }
+        />
       </Main>
     </Layout>
   );
